+++ src/Reflex/Dom/Contrib/Widgets/CheckboxList.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE FlexibleContexts         #-}
+{-# LANGUAGE GADTs                    #-}
+{-# LANGUAGE OverloadedStrings        #-}
 {-# LANGUAGE ScopedTypeVariables      #-}
 
 module Reflex.Dom.Contrib.Widgets.CheckboxList where
@@ -6,6 +9,7 @@ module Reflex.Dom.Contrib.Widgets.CheckboxList where
 import           Control.Monad
 import           Data.Set (Set)
 import qualified Data.Set as S
+import           Data.Text (Text)
 import           Reflex
 import           Reflex.Dom
 import           Reflex.Dom.Contrib.Widgets.Common
@@ -16,16 +20,16 @@ import           Reflex.Dom.Contrib.Widgets.Common
 -- | Takes a list of labels to make checkboxes for and returns the labels of
 -- the boxes that are checked.
 checkboxList
-    :: forall t m a. (MonadWidget t m, Ord a, Show a)
-    => (a -> String)
+    :: forall t m a. (MonadWidget t m, Ord a)
+    => (a -> Text)
     -- ^ Function to show each item
-    -> (String -> a -> Bool)
+    -> (Text -> a -> Bool)
     -- ^ Function to filter each item
     -> Event t Bool
     -- ^ Blanket event to apply to all list items.  Allows you to have "select
     -- all" and "select none" buttons.  Fire True to select all and False to
     -- select none.
-    -> Dynamic t String
+    -> Dynamic t Text
     -- ^ A search string for filtering the list of items.
     -> Set a
     -- ^ Set of items that should be initially checked
@@ -41,32 +45,32 @@ checkboxList showFunc filterFunc blanketEvent searchString onItems items = do
               if filterFunc search item
                 then mempty
                 else "style" =: "display:none"
-        attrs <- liftM nubDyn $ mapDyn mkAttrs searchString
+            attrs = uniqDyn $ mkAttrs <$> searchString
         elDynAttr "li" attrs $ el "label" $ do
           cb <- htmlCheckbox $ WidgetConfig
                   (leftmost [blanketEvent])
                   (S.member item onItems)
                   (constDyn mempty)
           text shown
-          mapWidget (\b -> if b then [item] else []) cb
-      wconcat es
+          return $ mapWidget (\b -> if b then [item] else []) cb
+      return $ wconcat es
 
 
 ------------------------------------------------------------------------------
 -- | Takes a list of labels to make checkboxes for and returns the labels of
 -- the boxes that are checked.
 checkboxListView
-    :: forall t m a b. (MonadWidget t m, Ord a, Show a)
-    => (a -> String)
+    :: forall t m a b. (MonadWidget t m, Ord a)
+    => (a -> Text)
     -- ^ Function to show each item
-    -> (String -> a -> Bool)
+    -> (Text -> a -> Bool)
     -- ^ Function to filter each item
     -> (a -> Bool -> b)
     -> Event t Bool
     -- ^ Blanket event to apply to all list items.  Allows you to have "select
     -- all" and "select none" buttons.  Fire True to select all and False to
     -- select none.
-    -> Dynamic t String
+    -> Dynamic t Text
     -- ^ A search string for filtering the list of items.
     -> Set a
     -- ^ Set of items that should be initially checked
@@ -83,7 +87,7 @@ checkboxListView showFunc filterFunc updateFunc blanketEvent searchString
               if filterFunc search item
                 then mempty
                 else "style" =: "display:none"
-        attrs <- liftM nubDyn $ mapDyn mkAttrs searchString
+        let attrs = uniqDyn $ mkAttrs <$> searchString
         elDynAttr "li" attrs $ el "label" $ do
           cb <- htmlCheckbox $ WidgetConfig
                   (leftmost [blanketEvent])