+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.5
+
+* Add route, fullUriRoute, partialPathRoute
+* Change ButtonGroup so radio button label is clickable.  This slightly
+  changes the markup structure, eliminating a <td> tag.
+
 # 0.4.2
 
 * Add DynamicList widget